Split tests into the main part and compat part
This is necessary because `pytest.importorskip` skips the entire test
file, and it's more readable.

Also move parametrized test name generation into a hook in
`conftest.py`.
