+++ tests/conftest.py
@@ -1,0 +1,15 @@
+import numpy
+
+
+def _is_shape(value):
+    """Return True if the value is a tuple of integers."""
+    return isinstance(value, tuple) and all(map(int.__instancecheck__, value))
+
+
+def pytest_make_parametrize_id(config, val, argname):
+    # Make names of parametrized tests prettier for certain arguments.
+    if argname == "shape" and _is_shape(val):
+        return "x".join(map(str, val))
+    if argname == "dtype" and isinstance(val, numpy.dtype):
+        return val.name
+    return None
